Refactor Product model and update database context
- Added using directive for eCommerce.Models in ProductDBContext.cs.
- Reintroduced Product class in Product.cs with updated properties.
- Changed Price property type from double to decimal for accuracy.
- Minor formatting adjustments in appsettings.json.
- Created/modified migration files to reflect database schema changes.
- Enhanced code clarity and maintainability with XML documentation.
