+++ eCommerce/Models/Product.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace eCommerce;
+namespace eCommerce.Models;
 
 /// <summary>
 /// An individual product for sale in the game store
@@ -23,5 +23,5 @@ public class Product
     /// The sales price of the product in USD
     /// </summary>
     [Range(0, 10_000, ErrorMessage = "Price must be between $0 and $10,000.")]
-    public double Price { get; set; } // Required, double, $0 - $10,000
+    public decimal Price { get; set; } // Required, double, $0 - $10,000
 }
