+++ eCommerce/appsettings.json
@@ -1,7 +1,7 @@
 {
   "ConnectionStrings": {
-        "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=GameStoreDB;Trusted_Connection=True"
-  }
+    "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=GameStoreDB;Trusted_Connection=True"
+  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
